Fix clang -Wshorten-64-to-32 warnings in port/ subdir (#6218)
git-svn-id: https://svn.osgeo.org/gdal/trunk@31482 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_base64.cpp
@@ -143,7 +143,7 @@ int CPLBase64DecodeInPlace(GByte* pszBase64)
                     break;
             }
         }
-        return(p-pszBase64);
+        return static_cast<int>(p-pszBase64);
     }
     return 0;
 }
